backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dj-database-url==3.1.2
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
icrawler==0.6.10
idna==3.11
jiter==0.13.0
lxml==6.0.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.2
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
razorpay==2.0.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
